+++ src/Event/Radio/AnnotateNextSong.php
@@ -21,6 +21,8 @@ final class AnnotateNextSong extends Event
 {
     private ?string $songPath;
 
+    private ?string $protocol = null;
+
     /** @var array Custom annotations that should be sent along with the AutoDJ response. */
     private array $annotations = [];
 
@@ -74,6 +76,11 @@ final class AnnotateNextSong extends Event
         $this->songPath = $songPath;
     }
 
+    public function setProtocol(string $protocol): void
+    {
+        $this->protocol = $protocol;
+    }
+
     public function isAsAutoDj(): bool
     {
         return $this->asAutoDj;
@@ -96,7 +103,17 @@ final class AnnotateNextSong extends Event
                 $annotationsStr[] = $annotationKey . '="' . $annotationVal . '"';
             }
 
-            return 'annotate:' . implode(',', $annotationsStr) . ':' . $this->songPath;
+            $annotateParts = [
+                'annotate',
+                implode(',', $annotationsStr),
+                $this->songPath
+            ];
+
+            if (null !== $this->protocol) {
+                array_unshift($annotateParts, $this->protocol);
+            }
+
+            return implode(':', $annotateParts);
         }
 
         return $this->songPath;