+++ src/Radio/AutoDJ/Annotations.php
@@ -36,6 +36,7 @@ final class Annotations implements EventSubscriberInterface
                 ['annotateForLiquidsoap', 15],
                 ['annotatePlaylist', 10],
                 ['annotateRequest', 5],
+                ['enableAutoCue', -5],
                 ['postAnnotation', -10],
             ],
         ];
@@ -180,6 +181,13 @@ final class Annotations implements EventSubscriberInterface
             $event->addAnnotations([
                 'request_id' => $request->getId(),
             ]);
+        }
+    }
+
+    public function enableAutoCue(AnnotateNextSong $event): void
+    {
+        if ($event->getStation()->getBackendConfig()->getEnableAutoCue()) {
+            $event->setProtocol('autocue');
         }
     }